rename : gallery, image repository  기능별 하위 패키지로 분리
