+++ src/main/java/com/template/server/domain/image/repository/comment/ImageCommentRepository.java
