+++ src/main/java/com/template/server/domain/image/repository/like/ImageLikeRepository.java
