+++ src/main/java/com/template/server/domain/image/repository/report/ImageReportRepository.java
