+++ src/main/java/com/template/server/domain/image/service/ImageViewService.java
@@ -3,7 +3,7 @@ package com.template.server.domain.image.service;
 import com.template.server.domain.image.entity.Gallery;
 import com.template.server.domain.image.entity.ImageView;
 import com.template.server.domain.image.repository.gallery.GalleryRepository;
-import com.template.server.domain.image.repository.ImageViewRepository;
+import com.template.server.domain.image.repository.view.ImageViewRepository;
 import com.template.server.domain.member.entity.Member;
 import com.template.server.domain.member.repository.MemberRepository;
 import com.template.server.global.error.exception.BusinessLogicException;
